+++ installer/__main__.py
@@ -1,7 +1,7 @@
 import traceback
 try:
     import installer
-except e:
+except:
     print("\n\n")
     print("An unexpected error occurred. Please open an issue on github!")
     print("here is the error:")
